+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Braintree Android SDK Release Notes
 
+
+## unreleased
+* PayPalNativeCheckout (BETA)
+  * Pass the risk correlation ID from the Native Checkout SDK if it is not provided in the initial PayPal request
+  
 ## 4.20.0
 
 * SharedUtils